# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cabinet CRM — façade de services",
    "description": "Façade HTTP localhost réutilisant le moteur Python du CRM. Toutes les routes /api/* exigent l'en-tête Authorization: Bearer <jeton>.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/patients": {
      "get": {
        "tags": ["patients"],
        "summary": "Patients List",
        "operationId": "patients_list_api_patients_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "", "title": "Search" }
          },
          {
            "name": "filtre",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "tous", "title": "Filtre" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Offset" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PatientListOut" } }
            }
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["patients"],
        "summary": "Patients Create",
        "operationId": "patients_create_api_patients_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PatientIn" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PatientOut" } }
            }
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/DuplicateWarning" } }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/patients/{patient_id}": {
      "get": {
        "tags": ["patients"],
        "summary": "Patient Detail",
        "operationId": "patient_detail_api_patients__patient_id__get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Patient Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PatientDetailOut" } }
            }
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["patients"],
        "summary": "Patients Update",
        "operationId": "patients_update_api_patients__patient_id__put",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Patient Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PatientIn" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PatientOut" } }
            }
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/patients/match": {
      "post": {
        "tags": ["patients"],
        "summary": "Patients Match",
        "operationId": "patients_match_api_patients_match_post",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PatientIn" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/crm__routers__patients__MatchOut" }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DuplicateWarning": {
        "properties": {
          "duplicate": { "type": "boolean", "title": "Duplicate" },
          "match": { "anyOf": [{ "$ref": "#/components/schemas/PatientOut" }, { "type": "null" }] }
        },
        "type": "object",
        "required": ["duplicate"],
        "title": "DuplicateWarning"
      },
...
      "ErrorBody": {
        "properties": {
          "code": { "type": "string", "title": "Code" },
          "message": { "type": "string", "title": "Message" }
        },
        "type": "object",
        "required": ["code", "message"],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": { "error": { "$ref": "#/components/schemas/ErrorBody" } },
        "type": "object",
        "required": ["error"],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatientDetailOut": {
        "properties": {
          "patient": { "$ref": "#/components/schemas/PatientOut" },
          "solde": { "$ref": "#/components/schemas/SoldeOut" },
          "nb_documents": { "type": "integer", "title": "Nb Documents" },
          "nb_encaissements": { "type": "integer", "title": "Nb Encaissements" },
          "has_sent": { "type": "boolean", "title": "Has Sent" },
          "nb_actes_a_regler": { "type": "integer", "title": "Nb Actes A Regler" },
          "total_a_regler": { "type": "number", "title": "Total A Regler" }
        },
        "type": "object",
        "required": [
          "patient",
          "solde",
          "nb_documents",
          "nb_encaissements",
          "has_sent",
          "nb_actes_a_regler",
          "total_a_regler"
        ],
        "title": "PatientDetailOut"
      },
      "PatientIn": {
        "properties": {
          "nom": { "type": "string", "title": "Nom" },
          "prenom": { "type": "string", "title": "Prenom" },
          "date_naissance": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Date Naissance"
          },
          "email": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Email" },
          "telephone": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Telephone"
          },
          "adresse": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Adresse" },
          "notes": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Notes" },
          "force": { "type": "boolean", "title": "Force", "default": false }
        },
        "type": "object",
        "required": ["nom", "prenom"],
        "title": "PatientIn"
      },
      "PatientListOut": {
        "properties": {
          "items": {
            "items": { "$ref": "#/components/schemas/PatientOut" },
            "type": "array",
            "title": "Items"
          },
          "total": { "type": "integer", "title": "Total" }
        },
        "type": "object",
        "required": ["items", "total"],
        "title": "PatientListOut"
      },
      "PatientOut": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "nom": { "type": "string", "title": "Nom" },
          "prenom": { "type": "string", "title": "Prenom" },
          "display": { "type": "string", "title": "Display" },
          "date_naissance": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Date Naissance"
          },
          "email": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Email" },
          "telephone": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Telephone"
          },
          "adresse": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Adresse" },
          "notes": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Notes" }
        },
        "type": "object",
        "required": ["id", "nom", "prenom", "display"],
        "title": "PatientOut"
      },
...
      "SoldeOut": {
        "properties": {
          "du": { "type": "number", "title": "Du" },
          "encaisse": { "type": "number", "title": "Encaisse" },
          "reste": { "type": "number", "title": "Reste" }
        },
        "type": "object",
        "required": ["du", "encaisse", "reste"],
        "title": "SoldeOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "crm__routers__patients__MatchOut": {
        "properties": {
          "matches": {
            "items": { "$ref": "#/components/schemas/PatientOut" },
            "type": "array",
            "title": "Matches"
          }
        },
        "type": "object",
        "required": ["matches"],
        "title": "MatchOut"
      },
...
    }
  }
}